Move more things in appshell_versions
