+++ appshell/client_app_gtk.cpp
@@ -83,23 +83,4 @@ double ClientApp::GetElapsedMilliseconds()
     return (time(NULL) - g_appStartupTime);
 }
 
-CefString ClientApp::AppGetSupportDirectory() 
-{
-    gchar *supportDir = g_strdup_printf("%s/%s", g_get_user_config_dir(), APP_NAME);
-    CefString result = CefString(supportDir);
-    g_free(supportDir);
-    
-    return result;
-}
-
-CefString ClientApp::AppGetDocumentsDirectory() 
-{
-    const char *dir = g_get_user_special_dir(G_USER_DIRECTORY_DOCUMENTS);
-    if (dir == NULL)  {
-        return AppGetSupportDirectory();
-    } else {
-        return CefString(dir);
-    }
-}
-
 }  // namespace client
