+++ appshell/client_app_mac.mm
@@ -86,17 +86,4 @@ std::string ClientApp::GetExtensionJSSource()
     return result;
 }
 
-
-CefString ClientApp::AppGetSupportDirectory() {
-  NSString *libraryDirectory = [NSSearchPathForDirectoriesInDomains(NSApplicationSupportDirectory, NSUserDomainMask, YES) objectAtIndex:0];
-  NSString *supportDirectory = [NSString stringWithFormat:@"%@/%@%@", libraryDirectory, GROUP_NAME, APP_NAME];
-  
-  return CefString([supportDirectory UTF8String]);
-}
-
-CefString ClientApp::AppGetDocumentsDirectory() {
-    NSString *documentsDirectory = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) objectAtIndex:0];
-    return CefString([documentsDirectory UTF8String]);
-}
-    
 }  // namespace client
