+++ appshell/client_app_win.cpp
@@ -38,28 +38,28 @@ namespace client {
 
 CefString ClientApp::GetCurrentLanguage()
 {
-	// Get the user's selected language
-	// Defaults to the system installed language if not using MUI.
-	LANGID langID = GetUserDefaultUILanguage();
+    // Get the user's selected language
+    // Defaults to the system installed language if not using MUI.
+    LANGID langID = GetUserDefaultUILanguage();
 
-	// Convert LANGID to a RFC 4646 language tag (per navigator.language)
-	int langSize = GetLocaleInfo(langID, LOCALE_SISO639LANGNAME, NULL, 0);
-	int countrySize = GetLocaleInfo(langID, LOCALE_SISO3166CTRYNAME, NULL, 0);
+    // Convert LANGID to a RFC 4646 language tag (per navigator.language)
+    int langSize = GetLocaleInfo(langID, LOCALE_SISO639LANGNAME, NULL, 0);
+    int countrySize = GetLocaleInfo(langID, LOCALE_SISO3166CTRYNAME, NULL, 0);
 
-	wchar_t *lang = new wchar_t[langSize + countrySize + 1];
-	wchar_t *country = new wchar_t[countrySize];
-	
-	GetLocaleInfo(langID, LOCALE_SISO639LANGNAME, lang, langSize);
-	GetLocaleInfo(langID, LOCALE_SISO3166CTRYNAME, country, countrySize);
+    wchar_t *lang = new wchar_t[langSize + countrySize + 1];
+    wchar_t *country = new wchar_t[countrySize];
 
-	// add hyphen
-	wcscat(wcscat(lang, L"-"), country);
-	std::wstring locale(lang);
+    GetLocaleInfo(langID, LOCALE_SISO639LANGNAME, lang, langSize);
+    GetLocaleInfo(langID, LOCALE_SISO3166CTRYNAME, country, countrySize);
 
-	delete [] lang;
-	delete [] country;
+    // add hyphen
+    wcscat(wcscat(lang, L"-"), country);
+    std::wstring locale(lang);
 
-	return CefString(locale);
+    delete [] lang;
+    delete [] country;
+
+    return CefString(locale);
 }
 
 std::string ClientApp::GetExtensionJSSource()
@@ -91,33 +91,6 @@ std::string ClientApp::GetExtensionJSSource()
 double ClientApp::GetElapsedMilliseconds()
 {
     return (timeGetTime() - g_appStartupTime);
-}
-
-CefString ClientApp::AppGetSupportDirectory() 
-{
-    wchar_t dataPath[MAX_UNC_PATH];
-    SHGetFolderPath(NULL, CSIDL_APPDATA, NULL, SHGFP_TYPE_CURRENT, dataPath);
-  
-    std::wstring appSupportPath = dataPath;
-    appSupportPath +=  L"\\" GROUP_NAME APP_NAME;
-
-    // Convert '\\' to '/'
-    replace(appSupportPath.begin(), appSupportPath.end(), '\\', '/');
-
-    return CefString(appSupportPath);
-}
-
-
-CefString ClientApp::AppGetDocumentsDirectory()
-{
-    wchar_t dataPath[MAX_UNC_PATH] = {0};
-    SHGetFolderPath(NULL, CSIDL_MYDOCUMENTS, NULL, SHGFP_TYPE_CURRENT, dataPath);
-    std::wstring appUserDocuments = dataPath;
-
-    // Convert '\\' to '/'
-    replace(appUserDocuments.begin(), appUserDocuments.end(), '\\', '/');
-
-    return CefString(appUserDocuments);
 }
 
 }  // namespace client
